move process_script and run_script_standalone out of cli.rs
